LF-7631: caam-decrypt: Taken care of extra spaces in decrypted data
Issue:-

Padding is not handled properly. At the time of decryption with caam-
decrypt, it prints extra spaces(padding) with the decrypted data.

Resolution:-

In PKCS7 padding style, last byte is reserved for no of bytes needed to be
padded. This solution trims the padding part from the decrypted data
using the last byte.

Result:-

No extra spaces in decrypted data.

Signed-off-by: Kshitiz Varshney <kshitiz.varshney@nxp.com>